requirements/pyproject.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=test --output-file=requirements/pyproject.txt --resolver=backtracking pyproject.toml
#
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via adsctl (pyproject.toml)
coverage[toml]==7.2.3
    # via
    #   adsctl (pyproject.toml)
    #   pytest-cov
exceptiongroup==1.1.1
    # via pytest
google-ads==21.0.0
    # via adsctl (pyproject.toml)
google-api-core==2.11.0
    # via google-ads
google-auth==2.17.3
    # via
    #   google-api-core
    #   google-auth-oauthlib
google-auth-oauthlib==1.0.0
    # via google-ads
googleapis-common-protos==1.59.0
    # via
    #   google-ads
    #   google-api-core
    #   grpcio-status
grpcio==1.54.0
    # via
    #   google-ads
    #   grpcio-status
grpcio-status==1.54.0
    # via google-ads
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via adsctl (pyproject.toml)
markupsafe==2.1.2
    # via jinja2
numpy==1.24.2
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via pytest
pandas==2.0.0
    # via adsctl (pyproject.toml)
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via adsctl (pyproject.toml)
proto-plus==1.22.2
    # via google-ads
protobuf==4.22.3
    # via
    #   google-ads
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.7
    # via adsctl (pyproject.toml)
pytest==7.3.1
    # via
    #   adsctl (pyproject.toml)
    #   pytest-cov
pytest-cov==4.0.0
    # via adsctl (pyproject.toml)
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via google-ads
requests==2.28.2
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
tabulate==0.9.0
    # via adsctl (pyproject.toml)
tomli==2.0.1
    # via
    #   coverage
    #   pytest
tomli-w==1.0.0
    # via adsctl (pyproject.toml)
tomlkit==0.11.7
    # via adsctl (pyproject.toml)
typing-extensions==4.5.0
    # via pydantic
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
